+++ src/backend/core/tests/items/test_api_item_upload_ended.py
@@ -113,12 +113,10 @@ def test_api_item_upload_ended_success():
 
     with (
         mock.patch.object(malware_detection, "analyse_file") as mock_analyse_file,
-        mock.patch("core.api.viewsets.mirror_item") as mock_mirror_item,
     ):
         response = client.post(f"/api/v1.0/items/{item.id!s}/upload-ended/")
 
     mock_analyse_file.assert_called_once_with(item.file_key, item_id=item.id)
-    mock_mirror_item.assert_called_once_with(item)
     assert response.status_code == 200
 
     item.refresh_from_db()
@@ -142,12 +140,10 @@ def test_api_item_upload_ended_empty_file():
 
     with (
         mock.patch.object(malware_detection, "analyse_file") as mock_analyse_file,
-        mock.patch("core.api.viewsets.mirror_item") as mock_mirror_item,
     ):
         response = client.post(f"/api/v1.0/items/{item.id!s}/upload-ended/")
 
     mock_analyse_file.assert_called_once_with(item.file_key, item_id=item.id)
-    mock_mirror_item.assert_called_once_with(item)
     assert response.status_code == 200
 
     item.refresh_from_db()